fix: resolved comments on pull request
