+++ src/api/connection.ts
@@ -1,7 +1,7 @@
 import { apiRoutes } from '../config/apiRoutes';
 import { storageKeys } from '../config/CommonConstant';
 import { getHeaderConfigs } from '../config/GetHeaderConfigs';
-import { axiosGet } from '../services/apiRequests';
+import { axiosDelete, axiosGet } from '../services/apiRequests';
 import { getFromLocalStorage } from './Auth';
 export interface IConnectionListAPIParameter {
 	itemPerPage: number;
@@ -34,3 +34,24 @@ export const getConnectionsByOrg = async ({
 		return err?.message;
 	}
 };
+
+
+export const deleteConnectionRecords = async (
+) => {
+	const orgId = await getFromLocalStorage(storageKeys.ORG_ID);
+
+	const url = `${apiRoutes.organizations.root}/${orgId}${apiRoutes.organizations.deleteConnections}`;
+
+	const axiosPayload = {
+		url,
+		config: await getHeaderConfigs(),
+	};
+
+	try {
+		return await axiosDelete(axiosPayload);
+	} catch (error) {
+		const err = error as Error;
+		return err?.message;
+	}
+};
+
