+++ src/api/issuance.ts
@@ -3,7 +3,7 @@ import { storageKeys } from '../config/CommonConstant';
 import {
 	getHeaderConfigs
 } from '../config/GetHeaderConfigs';
-import { axiosGet, axiosPost } from '../services/apiRequests';
+import { axiosDelete, axiosGet, axiosPost } from '../services/apiRequests';
 import { getFromLocalStorage } from './Auth';
 import type { IConnectionListAPIParameter } from './connection';
 
@@ -77,6 +77,27 @@ export const issueOobEmailCredential = async (data: object) => {
 
 	try {
 		return await axiosPost(axiosPayload);
+	} catch (error) {
+		const err = error as Error;
+		return err?.message;
+	}
+};
+
+
+
+export const deleteIssuanceRecords = async (
+) => {
+	const orgId = await getFromLocalStorage(storageKeys.ORG_ID);
+
+	const url = `${apiRoutes.organizations.root}/${orgId}${apiRoutes.organizations.deleteIssaunce}`;
+
+	const axiosPayload = {
+		url,
+		config: await getHeaderConfigs(),
+	};
+
+	try {
+		return await axiosDelete(axiosPayload);
 	} catch (error) {
 		const err = error as Error;
 		return err?.message;
